pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_project"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Your name (or your organization/company/team)", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "dvc-gs>=3.0.2",
    "evidently>=0.7.20",
    "fastapi==0.115.6",
    "httpx>=0.28.1",
    "hydra-core>=1.3.0",
    "kagglehub>=0.2.0",
    "lightning>=2.6.0",
    "pathspec<1",
    "pillow>=10.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "torch==2.6.0",
    "torchvision>=0.19.0",
    "tqdm>=4.65.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.16.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python==1.12.2",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "locust>=2.43.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]
